requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==2.3.7
